Extract GPT types from `image_aggregator.cc`
Bug: b/436625775
